pdf_vectorization/requirements.txt
langchain
langchain-community
langchain-core==0.1.10
pypdf==3.17.1
chromadb==0.4.22
sentence-transformers==2.2.2
streamlit==1.31.1
python-dotenv==1.0.0
openai==1.12.0
tiktoken==0.5.2 
langchain-openai==0.2.1
